com.unity.testtools.codecoverage@1.2.0 ## [1.2.0] - 2022-08-01
### Fixes
- Ensure assemblies are removed from the Included Assemblies field if they no longer exist (case [1318668](https://issuetracker.unity3d.com/issues/code-coverage-the-included-assemblies-field-shows-assemblies-that-no-longer-exist)).
- Ensure hidden sequence points are ignored (case [1372305](https://issuetracker.unity3d.com/issues/class-which-derives-from-methodbase-causes-incorrect-sequence-points-to-be-generated-by-coverage-api)).

### Changes
- Updated Report Generator to version 5.0.4.
- Updated the UI of the [Code Coverage window](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CodeCoverageWindow.html) moving the action buttons into a toolbar at the top.
- Renamed *assemblyFilters* aliases in [batchmode](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CoverageBatchmode.html); `<user>` was renamed to `<assets>` and `<project>` was renamed to `<all>`.
- Replaced `pathStrippingPatterns` with `pathReplacePatterns` in [batchmode](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CoverageBatchmode.html). The `pathReplacePatterns` option allows stripping and replacing specific sections from the paths that are stored in the coverage results xml files.

See the [Upgrade guide](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/upgrade-guide.html) if upgrading to Code Coverage package version 1.2.

### Improvements
- The size of the coverage result files and the Code Coverage session duration have been optimized. At the start of the session a coverage xml result file is generated which includes all the lines but with zero coverage. The following coverage xml result files that are generated within a Code Coverage session include only the coverage data of the visited lines.
- Added Help IconButton in the toolbar in the [Code Coverage window](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CodeCoverageWindow.html).
- Updated the mechanic for opening the containing folder, change the location or reset to the default location for *Results Location* and *Report History Location*.
- Refactored the [Code Coverage window](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CodeCoverageWindow.html) UI to include a new *Report Options* section and removing the word 'Generate' from the options.
- Introduced new selection buttons under the *Included Assemblies* dropdown in the [Code Coverage window](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CodeCoverageWindow.html); use the *All* button to select all the assemblies in the project. Use the *Assets* button to select only the assemblies under the `Assets` folder. Use the *Packages* button to select only the Packages' assemblies. If searching, the buttons will apply only to the assemblies visible in the list.
- Updated [What's new](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/whats-new.html) and [Upgrade guide](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/upgrade-guide.html) pages in the documentation.
- Added [Using relative paths in path filters](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CoverageBatchmode.html#using-relative-paths-in-path-filters) section in documentation.
- Updated the editor and console logs; added information about the assembly and path filters, improved coverage session logs.
- Improved the progress bars for `Writing coverage results` and `Generating the report`.
- Added an icon for the [Code Coverage window](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CodeCoverageWindow.html).
- Updated documentation to match version 1.2.0.

### Features
- Added `Pause Recording` and `Resume Recording` buttons in the toolbar in the [Code Coverage window](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CodeCoverageWindow.html).
- Added `Log Verbosity Level` setting in the [Code Coverage window](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CodeCoverageWindow.html) which allows setting the verbosity level for the editor and console logs.
- Added `Additional Reports` option in the [Code Coverage window](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CodeCoverageWindow.html) which if checked [SonarQube](https://docs.sonarqube.org/latest/analysis/generic-test), [Cobertura](https://cobertura.github.io/cobertura) and [LCOV](https://github.com/linux-test-project/lcov) reports will be generated. Added `generateAdditionalReports` in *-coverageOptions* for [batchmode](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CoverageBatchmode.html).
- Added `Test Runner References` report option in the [Code Coverage window](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CodeCoverageWindow.html) which if checked includes test references to the generated coverage results and enables the [Coverage by test methods](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/HowToInterpretResults.html#coverage-by-test-methods) section in the HTML report, allowing you to see how each test contributes to the overall coverage. Added `generateTestReferences` in *-coverageOptions* for [batchmode](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CoverageBatchmode.html).
- Added `Auto Open Report` option in the [Code Coverage window](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CodeCoverageWindow.html) which if checked the coverage report will open automatically after it has been generated.
- Added `pathFiltersFromFile` in *-coverageOptions* for [batchmode](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CoverageBatchmode.html) which allows specifying an external file which contains a list of path filters. When this file contains relative paths, the `sourcePaths` option can be used to specify the source directories.
- Added `dontClear` in *-coverageOptions* for [batchmode](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CoverageBatchmode.html) which allows coverage results to be accumulated after every code coverage session. If not passed the results are cleared before a new session. For more information see [Generate combined report from EditMode and PlayMode tests](https://docs.unity3d.com/Packages/com.unity.testtools.codecoverage@1.2/manual/CoverageBatchmode.html#generate-combined-report-from-editmode-and-playmode-tests).
- When the `pathFilters` option or the `pathFiltersFromFile` option in *-coverageOptions* contains relative paths, the `sourcePaths` option can be used to specify the source directories.

+++ Editor/Utils/CoverageUtils.cs
@@ -212,60 +212,45 @@ namespace UnityEditor.TestTools.CodeCoverage.Utils
 
         public static int GetNumberOfFilesInFolder(string folderPath, string filePattern, SearchOption searchOption)
         {
-            if (folderPath == null)
-                return 0;
+            if (folderPath != null && Directory.Exists(folderPath))
+            {
+                string[] files = Directory.GetFiles(folderPath, filePattern, searchOption);
+                return files.Length;
+            }
 
-            string[] files = Directory.GetFiles(folderPath, filePattern, searchOption);
-            return files.Length;
+            return 0;
         }
 
         public static void ClearFolderIfExists(string folderPath, string filePattern)
         {
-            if (folderPath != null)
+            if (folderPath != null && Directory.Exists(folderPath))
             {
-                if (Directory.Exists(folderPath))
+                DirectoryInfo dirInfo = new DirectoryInfo(folderPath);
+
+                foreach (FileInfo file in dirInfo.GetFiles(filePattern))
                 {
-                    DirectoryInfo dirInfo = new DirectoryInfo(folderPath);
-
-                    foreach (FileInfo file in dirInfo.GetFiles(filePattern))
+                    try
                     {
-                        try
-                        {
-                            file.Delete();
-                        }
-                        catch (Exception)
-                        {
-                            ResultsLogger.Log(ResultID.Warning_FailedToDeleteFile, file.FullName);
-                        }
+                        file.Delete();
                     }
-
-                    foreach (DirectoryInfo dir in dirInfo.GetDirectories())
+                    catch (Exception)
                     {
-                        try
-                        {
-                            dir.Delete(true);
-                        }
-                        catch (Exception)
-                        {
-                            ResultsLogger.Log(ResultID.Warning_FailedToDeleteDir, dir.FullName);
-                        }
+                        ResultsLogger.Log(ResultID.Warning_FailedToDeleteFile, file.FullName);
                     }
                 }
-            } 
-        }
 
-        public static bool DoesFolderExistAndNotEmpty(string folderPath)
-        {
-            if (folderPath != null)
-            {
-                if (Directory.Exists(folderPath))
+                foreach (DirectoryInfo dir in dirInfo.GetDirectories())
                 {
-                    DirectoryInfo dirInfo = new DirectoryInfo(folderPath);
-                    return dirInfo.GetFiles().Length > 0 || dirInfo.GetDirectories().Length > 0;
+                    try
+                    {
+                        dir.Delete(true);
+                    }
+                    catch (Exception)
+                    {
+                        ResultsLogger.Log(ResultID.Warning_FailedToDeleteDir, dir.FullName);
+                    }
                 }
             }
-
-            return false;
         }
 
         public static bool IsValidFolder(string folderPath)
@@ -328,6 +313,19 @@ namespace UnityEditor.TestTools.CodeCoverage.Utils
             if (startEndConstrains)
                 regex.Append("$");
             return regex.ToString();
+        }
+
+        public static string[] GetFilteringLogParams(AssemblyFiltering assemblyFiltering, PathFiltering pathFiltering, string[] otherParams = null)
+        {
+            string[] logParams = { assemblyFiltering != null && assemblyFiltering.includedAssemblies.Length > 0 ? assemblyFiltering.includedAssemblies : "<Not specified>",
+                    assemblyFiltering != null && assemblyFiltering.excludedAssembliesNoDefault.Length > 0 ? assemblyFiltering.excludedAssembliesNoDefault : "<Not specified>",
+                    pathFiltering != null && pathFiltering.includedPaths.Length > 0 ? pathFiltering.includedPaths : "<Not specified>",
+                    pathFiltering != null && pathFiltering.excludedPaths.Length > 0 ? pathFiltering.excludedPaths : "<Not specified>" };
+
+            if (otherParams != null && otherParams.Length > 0)
+                logParams = otherParams.Concat(logParams).ToArray();
+
+            return logParams;
         }
 
         [ExcludeFromCoverage]
